--- docs/src/pages/docs/api/canBe.en.mdx
@@ -1,105 +1,0 @@
----
-title: canBe
----
-
-import { Sandpack } from '@/components/Sandpack';
-
-# canBe
-
-## canBeChoseong
-
-Check if a given character can be a choseong in Korean.
-
-```typescript
-function canBeChoseong(character: string): boolean;
-```
-
-### Examples
-
-```typescript
-canBeChoseong('ㄱ'); // true
-canBeChoseong('ㅃ'); // true
-canBeChoseong('ㄱㅅ'); // false
-canBeChoseong('ㅏ'); // false
-canBeChoseong('가'); // false
-```
-
-### Demo
-
-<br />
-
-<Sandpack>
-
-```ts index.ts
-import { canBeChoseong } from 'es-hangul';
-
-console.log(canBeChoseong('ㄱ'));
-```
-
-</Sandpack>
-
-## canBeJungseong
-
-Check if a given character can be a jungseong in Korean.
-
-```typescript
-function canBeJungseong(character: string): boolean;
-```
-
-### Examples
-
-```typescript
-canBeJungseong('ㅏ'); // true
-canBeJungseong('ㅗㅏ'); // true
-canBeJungseong('ㅏㅗ'); // false
-canBeJungseong('ㄱ'); // false
-canBeJungseong('ㄱㅅ'); // false
-canBeJungseong('가'); // false
-```
-
-### Demo
-
-<br />
-
-<Sandpack>
-
-```ts index.ts
-import { canBeJungseong } from 'es-hangul';
-
-console.log(canBeJungseong('ㅏ'));
-```
-
-</Sandpack>
-
-## canBeJongseong
-
-Check if a given character can be a jongseong in Korean.
-
-```typescript
-function canBeJongseong(character: string): boolean;
-```
-
-### Examples
-
-```typescript
-canBeJongseong('ㄱ'); // true
-canBeJongseong('ㄱㅅ'); // true
-canBeJongseong('ㅎㄹ'); // false
-canBeJongseong('가'); // false
-canBeJongseong('ㅏ'); // false
-canBeJongseong('ㅗㅏ'); // false
-```
-
-### Demo
-
-<br />
-
-<Sandpack>
-
-```ts index.ts
-import { canBeJongseong } from 'es-hangul';
-
-console.log(canBeJongseong('ㄱ'));
-```
-
-</Sandpack>
