--- docs/src/pages/docs/api/canBe.ko.mdx
@@ -1,105 +1,0 @@
----
-title: canBe
----
-
-import { Sandpack } from '@/components/Sandpack';
-
-# canBe
-
-## canBeChoseong
-
-인자로 받은 문자가 초성으로 위치할 수 있는 문자인지 검사합니다.
-
-```typescript
-function canBeChoseong(character: string): boolean;
-```
-
-### Examples
-
-```typescript
-canBeChoseong('ㄱ'); // true
-canBeChoseong('ㅃ'); // true
-canBeChoseong('ㄱㅅ'); // false
-canBeChoseong('ㅏ'); // false
-canBeChoseong('가'); // false
-```
-
-### 사용해보기
-
-<br />
-
-<Sandpack>
-
-```ts index.ts
-import { canBeChoseong } from 'es-hangul';
-
-console.log(canBeChoseong('ㅏ'));
-```
-
-</Sandpack>
-
-## canBeJungseong
-
-인자로 받은 문자가 중성으로 위치할 수 있는 문자인지 검사합니다.
-
-```typescript
-function canBeJungseong(character: string): boolean;
-```
-
-### Examples
-
-```typescript
-canBeJungseong('ㅏ'); // true
-canBeJungseong('ㅗㅏ'); // true
-canBeJungseong('ㅏㅗ'); // false
-canBeJungseong('ㄱ'); // false
-canBeJungseong('ㄱㅅ'); // false
-canBeJungseong('가'); // false
-```
-
-### 사용해보기
-
-<br />
-
-<Sandpack>
-
-```ts index.ts
-import { canBeJungseong } from 'es-hangul';
-
-console.log(canBeJungseong('ㅏ'));
-```
-
-</Sandpack>
-
-## canBeJongseong
-
-인자로 받은 문자가 종성으로 위치할 수 있는 문자인지 검사합니다.
-
-```typescript
-function canBeJongseong(character: string): boolean;
-```
-
-### Examples
-
-```typescript
-canBeJongseong('ㄱ'); // true
-canBeJongseong('ㄱㅅ'); // true
-canBeJongseong('ㅎㄹ'); // false
-canBeJongseong('가'); // false
-canBeJongseong('ㅏ'); // false
-canBeJongseong('ㅗㅏ'); // false
-```
-
-### 사용해보기
-
-<br />
-
-<Sandpack>
-
-```ts index.ts
-import { canBeJongseong } from 'es-hangul';
-
-console.log(canBeJongseong('ㄱ'));
-```
-
-</Sandpack>
