+++ docs/src/pages/docs/api/canBeChoseong.ko.mdx
@@ -1,0 +1,37 @@
+---
+title: canBeChoseong
+---
+
+import { Sandpack } from '@/components/Sandpack';
+
+# canBeChoseong
+
+인자로 받은 문자가 초성으로 위치할 수 있는 문자인지 검사합니다.
+
+```typescript
+function canBeChoseong(character: string): boolean;
+```
+
+## Examples
+
+```typescript
+canBeChoseong('ㄱ'); // true
+canBeChoseong('ㅃ'); // true
+canBeChoseong('ㄱㅅ'); // false
+canBeChoseong('ㅏ'); // false
+canBeChoseong('가'); // false
+```
+
+## 사용해보기
+
+<br />
+
+<Sandpack>
+
+```ts index.ts
+import { canBeChoseong } from 'es-hangul';
+
+console.log(canBeChoseong('ㅏ'));
+```
+
+</Sandpack>
