+++ docs/src/pages/docs/api/canBeJongseong.ko.mdx
@@ -1,0 +1,38 @@
+---
+title: canBeJongseong
+---
+
+import { Sandpack } from '@/components/Sandpack';
+
+# canBeJongseong
+
+인자로 받은 문자가 종성으로 위치할 수 있는 문자인지 검사합니다.
+
+```typescript
+function canBeJongseong(character: string): boolean;
+```
+
+## Examples
+
+```typescript
+canBeJongseong('ㄱ'); // true
+canBeJongseong('ㄱㅅ'); // true
+canBeJongseong('ㅎㄹ'); // false
+canBeJongseong('가'); // false
+canBeJongseong('ㅏ'); // false
+canBeJongseong('ㅗㅏ'); // false
+```
+
+## 사용해보기
+
+<br />
+
+<Sandpack>
+
+```ts index.ts
+import { canBeJongseong } from 'es-hangul';
+
+console.log(canBeJongseong('ㄱ'));
+```
+
+</Sandpack>
