+++ src/index.ts
@@ -1,17 +1,19 @@
+export { amountToHangul } from './amountToHangul';
 export { assemble } from './assemble';
+export { canBeChoseong } from './canBeChoseong';
+export { canBeJongseong } from './canBeJongseong';
+export { canBeJungseong } from './canBeJungseong';
 export { combineCharacter, combineVowels } from './combineCharacter';
-export { convertQwertyToHangul, convertQwertyToAlphabet } from './convertQwertyToAlphabet';
+export { convertQwertyToAlphabet, convertQwertyToHangul } from './convertQwertyToAlphabet';
 export { days } from './days';
 export { disassemble, disassembleToGroups } from './disassemble';
 export { disassembleCompleteCharacter } from './disassembleCompleteCharacter';
+export { getChoseong } from './getChoseong';
+export { hasBatchim } from './hasBatchim';
 export { josa } from './josa';
+export { numberToHangul } from './numberToHangul';
+export { numberToHangulMixed } from './numberToHangulMixed';
 export { removeLastCharacter } from './removeLastCharacter';
 export { romanize } from './romanize';
 export { standardizePronunciation } from './standardizePronunciation';
 export { susa } from './susa';
-export { hasBatchim } from './hasBatchim';
-export { canBeChoseong, canBeJongseong, canBeJungseong } from './canBe';
-export { getChoseong } from './getChoseong';
-export { amountToHangul } from './amountToHangul';
-export { numberToHangul } from './numberToHangul';
-export { numberToHangulMixed } from './numberToHangulMixed';
