+++ src/romanize/romanize.ts
@@ -1,6 +1,7 @@
+import { canBeChoseong } from '@/canBeChoseong';
 import { isHangulCharacter } from '../_internal/hangul';
 import { assemble } from '../assemble';
-import { canBeChoseong } from '../canBe';
+
 import { disassembleCompleteCharacter } from '../disassembleCompleteCharacter';
 import { standardizePronunciation } from '../standardizePronunciation';
 import { 종성_알파벳_발음, 중성_알파벳_발음, 초성_알파벳_발음 } from './constants';
